src/scheduler/requirements.txt
pandas
openpyxl
SQLAlchemy
geoalchemy2
psycopg2-binary
python-dateutil
pytz
requests
tensorflow==2.17.0
numpy
asyncpg
scikit-learn
matplotlib
seaborn
pydot
graphviz
reverse_geocoder